fixed color in example other cleanup

+++ source/Color.h
@@ -2,10 +2,12 @@
 #ifndef COLOR_H_
 #define COLOR_H_
 
+#include <cstdint>
+
 struct Color {
-	int r;
-	int g;
-	int b;
+	uint16_t r;
+	uint16_t g;
+	uint16_t b;
 };
 
 #endif
